+++ components/ui/TabBarBackground.ios.tsx
@@ -1,16 +1,23 @@
 import { useBottomTabBarHeight } from '@react-navigation/bottom-tabs';
 import { BlurView } from 'expo-blur';
-import { StyleSheet } from 'react-native';
+import styled from 'styled-components/native';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
+
+const StyledBlurView = styled(BlurView)`
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+`;
 
 export default function BlurTabBarBackground() {
   return (
-    <BlurView
+    <StyledBlurView
       // System chrome material automatically adapts to the system's theme
       // and matches the native tab bar appearance on iOS.
       tint="systemChromeMaterial"
       intensity={100}
-      style={StyleSheet.absoluteFill}
     />
   );
 }